requirement.txt
streamlit==1.18.1
opencv-python-headless==4.5.5.64
Pillow==8.4.0
tensorflow==2.9.0
numpy==1.21.2
pandas==1.3.3
matplotlib==3.4.3
scikit-learn==0.24.2
requests==2.26.0
